+++ ChangeLog.md
@@ -3,6 +3,7 @@
 ## 0.3.0.0
 
 * Add support for foreign-lib components
+* Make `Sha256` type abstract and add new `sha256ToByteString` deconstructor
 
 ### Executable